+++ tests/bug-2794-opencode-model-profile-overrides.test.cjs
@@ -35,9 +35,8 @@ const {
   install,
 } = require('../bin/install.js');
 
-function makeTmp(prefix) {
-  return fs.mkdtempSync(path.join(os.tmpdir(), `gsd-2794-${prefix}-`));
-}
+const { createTempDir } = require('./helpers.cjs');
+const makeTmp = (prefix) => createTempDir(`gsd-2794-${prefix}-`);
 
 function writeJson(p, obj) {
   fs.mkdirSync(path.dirname(p), { recursive: true });
