+++ tests/bug-3231-false-gsd-sdk-ready-linux.test.cjs
@@ -36,6 +36,7 @@ const os = require('node:os');
 const path = require('node:path');
 
 const installModule = require('../bin/install.js');
+const { captureConsole } = require('./helpers.cjs');
 
 const isWindows = process.platform === 'win32';
 
@@ -49,33 +50,6 @@ const {
 // ---------------------------------------------------------------------------
 // Console capture helper (no ANSI)
 // ---------------------------------------------------------------------------
-function captureConsole(fn) {
-  const stdout = [];
-  const stderr = [];
-  const origLog = console.log;
-  const origWarn = console.warn;
-  const origError = console.error;
-  console.log = (...a) => stdout.push(a.join(' '));
-  console.warn = (...a) => stderr.push(a.join(' '));
-  console.error = (...a) => stderr.push(a.join(' '));
-  let threw = null;
-  try {
-    fn();
-  } catch (e) {
-    threw = e;
-  } finally {
-    console.log = origLog;
-    console.warn = origWarn;
-    console.error = origError;
-  }
-  if (threw) throw threw;
-  const strip = (s) => s.replace(/\x1b\[[0-9;]*m/g, '');
-  return {
-    stdout: stdout.map(strip).join('\n'),
-    stderr: stderr.map(strip).join('\n'),
-  };
-}
-
 // ---------------------------------------------------------------------------
 // Shared fixture helpers
 // ---------------------------------------------------------------------------
