+++ tests/helpers.cjs
@@ -175,4 +175,52 @@ function isUsageOutput(text) {
   return /Usage:\s*gsd-tools/.test(text) && /Commands:/.test(text);
 }
 
-module.exports = { runGsdTools, createTempDir, createTempProject, createTempGitProject, cleanup, parseFrontmatter, isUsageOutput, TOOLS_PATH };
+/**
+ * Run `fn` with console.log/warn/error captured, returning {stdout, stderr}
+ * with ANSI colors stripped. Re-throws any exception fn threw AFTER restoring
+ * the real console so the caller's assertion path sees the failure (without
+ * this, a fn that crashes before printing would falsely pass !hasReady-style
+ * assertions). #2775 CR follow-up established this exact contract.
+ *
+ * Previously duplicated in bug-2775, bug-2829, bug-3033, bug-3211, bug-3231,
+ * bug-3359, and installer-migration-install-integration.
+ */
+function captureConsole(fn) {
+  const stdout = [];
+  const stderr = [];
+  const origLog = console.log;
+  const origWarn = console.warn;
+  const origError = console.error;
+  console.log = (...a) => stdout.push(a.join(' '));
+  console.warn = (...a) => stderr.push(a.join(' '));
+  console.error = (...a) => stderr.push(a.join(' '));
+  let threw = null;
+  try {
+    fn();
+  } catch (e) {
+    threw = e;
+  } finally {
+    console.log = origLog;
+    console.warn = origWarn;
+    console.error = origError;
+  }
+  if (threw) throw threw;
+  const strip = (s) => s.replace(/\x1b\[[0-9;]*m/g, '');
+  return {
+    stdout: stdout.map(strip).join('\n'),
+    stderr: stderr.map(strip).join('\n'),
+  };
+}
+
+/**
+ * Normalize platform path separators to POSIX forward slashes. Use for
+ * cross-platform path comparisons in test assertions where the runtime
+ * emits the platform-native separator (\ on Windows) but the test
+ * fixture or expected literal is POSIX. Returns the input unchanged if
+ * null/undefined so it composes safely with optional chaining.
+ */
+function toPosixPath(p) {
+  return p == null ? p : p.split(path.sep).join('/');
+}
+
+module.exports = { runGsdTools, createTempDir, createTempProject, createTempGitProject, cleanup, parseFrontmatter, isUsageOutput, captureConsole, toPosixPath, TOOLS_PATH };
